requirements.txt
aniso8601==1.0.0
Flask==0.10.1
Flask-RESTful==0.3.4
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pytz==2015.6
six==1.10.0
Werkzeug==0.10.4
pycrypto
